Stop using unsigned types in database schema
Because unsigned integer types in .NET are not CLS compliant [1],
columns that previously had an unsigned integer type in the database
schema were mapped to properties with larger signed integer type in the
C# model classes.

Although I'm confident we could continue with this approach by
explicitly specifying column types in Entity Framework, I'd rather stick
to the default Entity Framework conventions wherever possible. This
commit therefore aligns the types in the database schema to match the
.NET types used in the model classes.

[1] https://learn.microsoft.com/en-us/dotnet/standard/language-independence#types-and-type-member-signatures

+++ src/Buttercup.DataAccess/PasswordResetTokenDataProvider.cs
@@ -35,7 +35,7 @@ internal sealed class PasswordResetTokenDataProvider : IPasswordResetTokenDataPr
         command.CommandText = "SELECT user_id FROM password_reset_token WHERE token = @token";
         command.Parameters.AddWithValue("@token", token);
 
-        return await command.ExecuteScalarAsync<uint?>();
+        return await command.ExecuteScalarAsync<long?>();
     }
 
     public async Task InsertToken(MySqlConnection connection, long userId, string token)
